feat: added unit tests for User_Service and Sale_Services.+

+++ src/main/java/com/firstproject/smartinventory/service/UserServiceImpl.java
@@ -2,6 +2,7 @@ package com.firstproject.smartinventory.service;
 
 import com.firstproject.smartinventory.dto.UserRequestDTO;
 import com.firstproject.smartinventory.dto.UserResponseDTO;
+import com.firstproject.smartinventory.entity.Store;
 import com.firstproject.smartinventory.entity.User;
 import com.firstproject.smartinventory.mapper.UserMapper;
 import com.firstproject.smartinventory.repository.UserRepository;
@@ -23,11 +24,21 @@ public class UserServiceImpl implements UserService{
     @Autowired
     private PasswordEncoder passwordEncoder;
 
+    @Autowired
+    private  StoreContextService storeContextService;
+
+    @Autowired
+    private StoreAuthorizationService storeAuthorizationService;
+
     @Override
     public UserResponseDTO createUser( UserRequestDTO userRequestDTO) {
-        if(userRepository.existsByUserNameIgnoreCase(userRequestDTO.getUserName())){
-            throw new RuntimeException("User name is already exist");
+        Store store = storeContextService.getCurrentStore();
+        storeAuthorizationService.verifyUserAccess(store);
+        if(userRepository.existsByUserNameIgnoreCaseAndStore(userRequestDTO.getUserName(),store)){
+            throw new RuntimeException("User name is already exist.");
         }
+        if(userRepository.existsByEmailAndStore(userRequestDTO.getEmail(), store))
+            throw new IllegalArgumentException("Email already exists in database.");
 
         User user = UserMapper.toEntity(userRequestDTO);
         user.setPassword(passwordEncoder.encode(userRequestDTO.getPassword()));
@@ -38,16 +49,22 @@ public class UserServiceImpl implements UserService{
 
     @Override
     public UserResponseDTO getUserById(String id) {
-        return userRepository.findById(id)
+
+        if(id == null)
+            throw new IllegalArgumentException("Entered ID is not valid.");
+        Store store = storeContextService.getCurrentStore();
+        storeAuthorizationService.verifyUserAccess(store);
+        return userRepository.findByUserIdAndStore(id,store)
                 .map(UserMapper::toDTO)
                 .orElseThrow(() ->
-                        new EntityNotFoundException("User not found "+ id));
+                        new IllegalArgumentException("User not found with ID "+ id));
     }
 
     @Override
     public List<UserResponseDTO> getAllUsers() {
-
-        return userRepository.findAll()
+        Store store = storeContextService.getCurrentStore();
+        storeAuthorizationService.verifyUserAccess(store);
+        return userRepository.findAllUsersByStore(store)
                 .stream()
                 .map(UserMapper::toDTO)
                 .collect(Collectors.toList());
@@ -55,26 +72,36 @@ public class UserServiceImpl implements UserService{
 
     @Override
     public UserResponseDTO updateUser(String id, UserRequestDTO userRequestDTO) {
-        User user = userRepository.findById(id)
-                .orElseThrow(()-> new UsernameNotFoundException("User is Not found "+id ));
+        if(id == null)
+            throw new IllegalArgumentException("Enter a valid userID");
 
-        if(userRequestDTO.getId() != null)
-            user.setId(userRequestDTO.getId());
+        Store store =storeContextService.getCurrentStore();
+        storeAuthorizationService.verifyUserAccess(store);
+
+        User user = userRepository.findByUserIdAndStore(id,store)
+                .orElseThrow(()-> new UsernameNotFoundException("User is not found "+id ));
+
         if(userRequestDTO.getUserName() != null)
             user.setUserName(userRequestDTO.getUserName());
         if(userRequestDTO.getRole() != null)
             user.setRole(userRequestDTO.getRole());
         if(userRequestDTO.getPassword() != null)
             user.setPassword(userRequestDTO.getPassword());
-
-        User updatedUser = userRepository.save(user);
-        return UserMapper.toDTO(updatedUser);
+        if(userRequestDTO.getEmail()!=null)
+            user.setEmail(userRequestDTO.getEmail());
+        if(userRequestDTO.getId()!=null){
+            user.setId(userRequestDTO.getId());
+        }
+        userRepository.save(user);
+        return UserMapper.toDTO(user);
     }
 
     @Override
     public void deleteUser(String id) {
-        User user = userRepository.findById(id)
-                .orElseThrow(()-> new UsernameNotFoundException("User is Not found "+id ));
+        Store store = storeContextService.getCurrentStore();
+        storeAuthorizationService.verifyUserAccess(store);
+        User user = userRepository.findByUserIdAndStore(id,store)
+                .orElseThrow(()-> new RuntimeException("User is not found "+id ));
         userRepository.delete(user);
     }
 
@@ -83,7 +110,7 @@ public class UserServiceImpl implements UserService{
         return userRepository.findAll().stream()
                 .filter(user -> user.getUserName().equalsIgnoreCase(name))
                 .findFirst()
-                .map(user -> new UserResponseDTO(user.getId(),user.getUserName(),user.getRole()))
+                .map(user -> new UserResponseDTO(user.getId(),user.getUserName(),user.getRole(), user.getEmail()))
                 .orElseThrow(()-> new UsernameNotFoundException("No user is found with "+name));
 
     }
